+++ uni-back/src/routes/posts/:id/incrementView.ts
@@ -1,0 +1,39 @@
+// import { number } from 'joi';
+import { Context } from 'koa';
+import Router from 'koa-router';
+
+import { IJWTState, jwtWithSetUserModel } from '../../../middlewares/jwt';
+
+import Posts from '../../../db/models/Posts.model';
+import { IPostPostLikeReq } from '../../../middlewares/joi-posts';
+
+export default function registerRoute(router: Router) {
+  interface PostPostIncrementView extends Context {
+    state: IJWTState;
+    request: IPostPostLikeReq;
+  }
+  router.post('/', jwtWithSetUserModel, async (ctx: PostPostIncrementView) => {
+    const postId: number = parseInt(ctx.params.id);
+    if (postId == null) {
+      ctx.throw(400, 'Id must be number');
+    }
+
+    const curPost = await Posts.findOne({
+      where: {
+        id: postId,
+      },
+    });
+
+    if (curPost == null) {
+      ctx.throw(404, 'Post not Found');
+    }
+
+    await curPost.increment('viewsCount', { silent: true });
+
+    ctx.status = 200;
+    ctx.type = 'json';
+    ctx.body = {
+      currentViewsCount: curPost.viewsCount,
+    };
+  });
+}
